requirements.txt
Django==3.2.20
Pillow==9.5.0